chore: filled `criteria.json` and `test_base.py` and `test_bonus.py` for local testing.
