+++ criteria.json
@@ -12,15 +12,7 @@
           "weight" : 50,
           "tests" : {
           "test_html_quantitative_div_tags": {
-            "checks": 30,
-            "weight": 20
-          },
-           "test_html_quantitative_div": {
-            "checks": 30,
-            "weight": 20
-          },
-           "test_html_div": {
-            "checks": 30,
+            "checks": 2,
             "weight": 20
           },
           "test_html_quantitative_section_tags" : {
@@ -29,15 +21,6 @@
           }
         }
         }
-      },
-      "css": {
-        "weight": 10,
-        "test_path": "test_css_",
-        "exclude": ["test_css_external_css_only","test_css_consistent_units"]
-      },
-      "js": {
-        "weight": 30,
-        "test_path": "test_js_"
       }
     }
   },
@@ -48,7 +31,16 @@
       "html": {
         "weight": 100,
         "test_path": "test_html_",
-        "include": ["test_html_article_or_aside_used"]
+
+        "quantitative": {
+          "weight": 20,
+          "tests": {
+            "test_html_quantitative_div_tags": {
+              "checks": 2,
+              "weight": 20
+            }
+          }
+        }
       }
     }
   },
